backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
ultralytics
openai
python-dotenv
